+++ ironelder/AndroidArchitecture/app/src/main/java/com/ironelder/androidarchitecture/data/repository/SearchDataRepository.kt
@@ -2,9 +2,13 @@ package com.ironelder.androidarchitecture.data.repository
 
 import com.ironelder.androidarchitecture.data.TotalModel
 import io.reactivex.Single
-import retrofit2.Call
 
 interface SearchDataRepository {
-    fun getDataForSearch(type: String, query: String): Call<TotalModel>
-    fun getDataForSearchWithAdapter(type: String, query: String?): Single<TotalModel>
+
+    fun getDataForSearch(
+        type: String,
+        query: String?,
+        get: (observable: Single<TotalModel>) -> Unit
+    )
+
 }
